pyproject.toml
[tool.poetry]
name = "gpt-sheets"
version = "0.1.0"
description = ""
authors = ["Hiro"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.6"
google-api-python-client = "^2.86.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.30.0"
python-dotenv = "^1.0.0"
gspread = "^5.8.0"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.23.0"
jupyterlab = "^3.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
